fixed issue of User Name not updating
